+++ app/src/main/java/br/com/hay/hay/Base/BasePresenter.kt
@@ -2,4 +2,5 @@ package br.com.hay.hay.Base
 
 interface BasePresenter {
    fun start()
+   fun finish()
 }
